Move more look related functions to `lib_looks`
